TEX - Reduce to one graphics file and use independet tripe buffers for graphics and oled
